+++ client/my-sites/stats/stats-redirect/load-stats-page.tsx
